remove empty-lines-after 1 in capture template
I seem to have an occasional issue with the org-capture templates. I can
usually fix them by going to the file and saving, but it gets annoying
after awhile. I think part of the reason for this issue is the space
between the =Personal= header and the =School= header. Hopefully, this
fixes the problem.